update code
Delete cart and update product quantity when creating new order, change the logic when creating a new whitelist.

+++ order-service/src/main/java/com/example/orderservice/dto/request/CartItemRequest.java
@@ -10,7 +10,7 @@ import java.math.BigDecimal;
 @Builder
 public class CartItemRequest {
     private Long productId;
-    private Integer userId; // to find shopping cart
+    private Long userId; // to find shopping cart
     private Integer quantity;
     private BigDecimal unitPrice;
 }
